backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
mysql-connector-python==8.0.33
PyJWT==2.6.0
stripe==5.4.0
python-decouple==3.8
gunicorn==20.1.0

